+++ lib/src/num/pd.rs
@@ -1,8 +1,8 @@
 #[derive(Clone, Copy, Debug)]
 pub struct Pd(pub usize, pub usize);
 impl Pd {
-    pub fn fr(&self) -> f64 {
-        self.0 as f64 / self.1 as f64
+    pub fn fr(&self) -> f32 {
+        self.0 as f32 / self.1 as f32
     }
     pub fn mul(&self, mul: usize) -> Self {
         Self(self.0 * mul, self.1)
